pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "papermind"
version = "1.1.0"
description = "Scientific knowledge base — papers, packages, codebases → queryable markdown"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "httpx>=0.27",
    "griffe>=1.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0",
    "rich>=13.0",
    "jinja2>=3.1",
    "mcp>=1.0",
]

[project.optional-dependencies]
ocr = ["transformers>=4.48", "torch>=2.0", "torchvision>=0.20", "accelerate>=1.0", "pymupdf>=1.24"]
browser = ["playwright>=1.40"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[project.scripts]
papermind = "papermind.cli.main:app"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
